backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.31.0
sqlparse==0.5.5
tzdata==2025.3
